requirements.txt
pygame==2.6.1
pyserial==3.5
setuptools==80.9.0
